+++ client/features/app/index.ts
@@ -1,0 +1,4 @@
+export { AuthGate } from "./AuthGate";
+export { OnboardingGate } from "./OnboardingGate";
+export { useAppStartup } from "./useAppStartup";
+export { isAuthEnabled } from "./isAuthEnabled";